+++ packages/keyring-snap-bridge/src/index.ts
@@ -1,5 +1,6 @@
 export * from './options';
 export * from './types';
 export * from './SnapKeyring';
+export * from './v2/SnapKeyringV1Adapter';
 export * from './SnapKeyringV1';
 export type * from './SnapKeyringMessenger';